requirements.txt
streamlit
tensorflow
ultralytics
pillow
numpy
easyocr
groq
segmentation-models-pytorch
opencv-python-headless
tzlocal
torchvision
sqlalchemy
phidata
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
duckduckgo-search
google-generativeai
opencv-python
gdown
torch
torchvision
segmentation-models-pytorch
